# misc/api-doc.yaml
...
openapi: 3.0.0
info:
  title: 專題 api
  description: |
    This is for project app api
  version: 1.0.0
servers:
- url: http://localhost:3000
...
paths:
...
  /comments/{postNum}:
    get:
      tags:
      - comment
      summary: 取得 留言
      description: 取得留言、留言者資訊
      operationId: getComments
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
              - memName: Fanny
                memPhoto: "1633331135718_2.jpg"
                memIdentify: 0
                commentNum: 4
                postNum: 2
                memNum: 2
                commentCreateAt: 2021-09-06T07:14:53.000Z
                commentUpdateAt: 2021-09-06T07:14:53.000Z
                commentContent: 什麼時候可以完成TT
              - memName: 野原新之助
                memPhoto: ""
                memIdentify: 1
                commentNum: 3
                postNum: 2
                memNum: 3
                commentCreateAt: 2021-09-06T07:10:40.000Z
                commentUpdateAt: 2021-09-06T07:10:40.000Z
                commentContent: 什麼時候可以完成TT
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...
  /comments/create/{postNum}:
    post:
      tags:
      - comment
      summary: 新增 留言
      description: RequestBody各項參數說明 <ul> <li>memNum：用戶序號</li> <li>commentContent：留言內容</li> </ul>
      operationId: createComment
      parameters:
      - name: postNum
        in: path
        description: 文章序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 2
              commentContent: 我也這麼覺得！
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 留言新增成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /comments/update/{commentNum}:
    put:
      tags:
      - comment
      summary: 編輯 留言
      description: RequestBody各項參數說明 <ul> <li>commentContent：留言內容</li> </ul>
      operationId: updateComment
      parameters:
      - name: commentNum
        in: path
        description: 留言序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              commentContent: 這個商品有毒嗎？
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 留言編輯成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /comments/delete/{commentNum}:
    delete:
      tags:
      - comment
      summary: 刪除 留言
      operationId: deleteComment
      parameters:
      - name: commentNum
        in: path
        description: 留言序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 留言刪除成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    post:
      tags:
      - comment
      summary: 留言按讚
      description: RequestBody各項參數說明 <ul> <li>memNum：按讚者序號</li> </ul>
      operationId: commentLike
      parameters:
      - name: commentNum
        in: path
        description: 留言序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功為留言按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /comments/liked/{commentNum}/{memNum}:
    delete:
      tags:
      - comment
      summary: 取消 留言按讚
      operationId: removeCommentLike
      parameters:
      - name: commentNum
        in: path
        description: 留言序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功取消貼文按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    get:
      tags:
      - comment
      summary: 檢查用戶是否按讚
      operationId: commentLikeVerify
      parameters:
      - name: commentNum
        in: path
        description: 留言序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 已按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /comments/report/{commentNum}:
    post:
      tags:
      - comment
      summary: 檢舉留言
      description: RequestBody各項參數說明 <ul> <li>memNum：檢舉者序號</li> </ul>
      operationId: commentReport
      parameters:
      - name: commentNum
        in: path
        description: 留言序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
              reportReason: 檢舉測試
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功檢舉留言
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...
  /members/{memNum}:
    get:
      tags:
      - member
      summary: 取得 個人資料
      operationId: getMemberInfo
      parameters:
      - name: memNum
        in: path
        description: 用戶序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                data:
                  memNum: 2
                  memAccount: fannyisfish
                  memPassword: t107A50034
                  memIdentify: 0
                  memPhoto: "1633331135718_2.jpg"
                  memName: Fanny
                  memIntro: Fanny test
                  memGender: 1
                  memBirth: 2000-05-05T16:00:00.000Z
                  likeNum: 0
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /members/update/{memNum}:
    put:
      tags:
      - member
      summary: 編輯 個人資料
      description: RequestBody各項參數說明 <ul> <li>memName：暱稱</li> <li>memIntro：簡介</li> <li>memGender：生理性別(0為男性，1為女性)</li> <li>memBirth：生日(格式：YYYY-MM-DD)</li> </ul> 身分為業務員才可編輯的欄位： <ul> <li>memCompany：所屬公司</li> <li>companyContact：公司聯絡方式</li> <li>memService：可提供服務</li> <li>memPhone：連絡資料-電話</li> <li>memLineID：連絡資料-LineID</li> </ul>
      operationId: updateMemberInfo
      parameters:
      - name: memNum
        in: path
        description: 用戶序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memName: 野原新之助
              memIntro: 測試測試
              memGender: 0
              memBirth: "1998-06-18"
              memCompany: "南山人壽"
              companyContact: "0911"
              memService: "壽險"
              memPhone: "0922"
              memLineID: "test123"
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 個人資料編輯成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /members/update-password/{memNum}:
    put:
      tags:
      - member
      summary: 變更 密碼
      operationId: updateMemPassword
      parameters:
      - name: memNum
        in: path
        description: 用戶序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memOldPassword: "test123"
              memNewPassword: "testqq"
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 變更密碼成功
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /members/liked/{salesmanNum}:
    post:
      tags:
      - member
      summary: 業務員按讚
      description: RequestBody各項參數說明 <ul> <li>memNum：按讚者序號</li> </ul>
      operationId: salesmanLike
      parameters:
      - name: salesmanNum
        in: path
        description: 被按讚(業務員)序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功為業務員按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /members/liked/{salesmanNum}/{memNum}:
    delete:
      tags:
      - member
      summary: 取消 業務員按讚
      operationId: removeSalesmantLike
      parameters:
      - name: salesmanNum
        in: path
        description: 被按讚(業務員)序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功取消業務員按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
    get:
      tags:
      - member
      summary: 檢查用戶是否按讚
      operationId: salesmanLikeVerify
      parameters:
      - name: salesmanNum
        in: path
        description: 被按讚(業務員)序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      - name: memNum
        in: path
        description: 按讚者序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 已按讚
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
  /members/report/{salesmanNum}:
    post:
      tags:
      - member
      summary: 檢舉業務員
      description: RequestBody各項參數說明 <ul> <li>memNum：檢舉者序號</li> </ul>
      operationId: salesmanReport
      parameters:
      - name: salesmanNum
        in: path
        description: 被檢舉者(業務員)序號
        required: true
        style: simple
        explode: false
        schema:
          type: number
          example: 3
      requestBody:
        content:
          application/json:
            schema: {}
            example:
              memNum: 1
              reportReason: 檢舉測試
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                message: 成功檢舉業務員
        "404":
          description: 找不到api
          content:
            application/json:
              example:
                message: Not found
        "500":
          description: Server 端錯誤
...